Extract chat completion and responses API logic into separate classes
- Create OpenAIChatCompletion class for chat API logic
- Create OpenAIResponses class for responses API logic
- Maintain all functionality including recursive tool handling
- OpenAIExecutor now delegates to these specialized classes
- Cleaner separation of concerns and easier to maintain
- Add delegator methods for logging from API handlers
